requirements.txt
paho-mqtt==1.3.0
pycrypto==2.6.1
web.py==0.40.dev0  
mysqlclient==1.3.12  
